backend-similarity-checker/requirements.txt
fastapi==0.95.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.5
uvicorn==0.22.0
sentence-transformers
faiss-cpu
numpy
